move ResourceForm to its own folder in components
